requirements.txt
python-dotenv==1.0.1
requests==2.32.3
python-chess==1.999
cairosvg==2.7.1
ipython==8.25.0
moviepy==1.0.3
openai==1.35.3
pip install -e .
